优化代码格式

+++ dmall-service-impl/dmall-service-impl-member/dmall-service-impl-member-business/src/main/java/com/dmall/mms/service/impl/memberinvoice/handler/DeleteMemberInvoiceHandler.java
@@ -3,6 +3,7 @@ package com.dmall.mms.service.impl.memberinvoice.handler;
 import com.dmall.common.dto.BaseResult;
 import com.dmall.common.util.ResultUtil;
 import com.dmall.component.web.handler.AbstractCommonHandler;
+import com.dmall.mms.api.enums.MmsErrorEnum;
 import com.dmall.mms.generator.dataobject.MemberInvoiceDO;
 import com.dmall.mms.generator.mapper.MemberInvoiceMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,6 +21,10 @@ public class DeleteMemberInvoiceHandler extends AbstractCommonHandler<Long, Memb
 
     @Override
     public BaseResult<Long> processor(Long id) {
+        MemberInvoiceDO memberInvoiceDO = memberInvoiceMapper.selectById(id);
+        if (memberInvoiceDO == null){
+            return ResultUtil.fail(MmsErrorEnum.INVOICE_NOT_EXISTS);
+        }
         memberInvoiceMapper.deleteById(id);
         return ResultUtil.success(id);
     }
